Test fixes
