+++ salt/minion.py
@@ -1048,7 +1048,30 @@ class MinionManager(MinionBase):
         #    self.opts,
         #    io_loop=self.io_loop,
         # )
+
+        # import hashlib
+        # ipc_publisher = salt.transport.publish_server(self.opts)
+        # hash_type = getattr(hashlib, self.opts["hash_type"])
+        # id_hash = hash_type(
+        #    salt.utils.stringutils.to_bytes(self.opts["id"])
+        # ).hexdigest()[:10]
+        # epub_sock_path = "ipc://{}".format(
+        #    os.path.join(
+        #        self.opts["sock_dir"], "minion_event_{}_pub.ipc".format(id_hash)
+        #    )
+        # )
+        # if os.path.exists(epub_sock_path):
+        #    os.unlink(epub_sock_path)
+        # epull_sock_path = "ipc://{}".format(
+        #    os.path.join(
+        #        self.opts["sock_dir"], "minion_event_{}_pull.ipc".format(id_hash)
+        #    )
+        # )
+        # ipc_publisher.pub_uri = epub_sock_path
+        # ipc_publisher.pull_uri = epull_sock_path
+        # self.io_loop.add_callback(ipc_publisher.publisher, ipc_publisher.publish_payload, self.io_loop)
         import hashlib
+
         ipc_publisher = salt.transport.publish_server(self.opts)
         hash_type = getattr(hashlib, self.opts["hash_type"])
         id_hash = hash_type(
@@ -1068,7 +1091,18 @@ class MinionManager(MinionBase):
         )
         ipc_publisher.pub_uri = epub_sock_path
         ipc_publisher.pull_uri = epull_sock_path
-        self.io_loop.add_callback(ipc_publisher.publisher, ipc_publisher.publish_payload, self.io_loop)
+        if self.opts["transport"] == "tcp":
+
+            def target():
+                ipc_publisher.publish_daemon(ipc_publisher.publish_payload)
+
+            proc = salt.utils.process.Process(target=target, daemon=True)
+            proc.start()
+        else:
+            self.io_loop.add_callback(
+                ipc_publisher.publisher, ipc_publisher.publish_payload, self.io_loop
+            )
+        log.error("get event ")
         self.event = salt.utils.event.get_event(
             "minion", opts=self.opts, io_loop=self.io_loop
         )
@@ -1139,11 +1173,8 @@ class MinionManager(MinionBase):
             self.io_loop.spawn_callback(self._connect_minion, minion)
         self.io_loop.call_later(timeout, self._check_minions)
 
-    @tornado.gen.coroutine
-    def _connect_minion(self, minion):
-        """
-        Create a minion, and asynchronously connect it to a master
-        """
+    async def _connect_minion(self, minion):
+        """Create a minion, and asynchronously connect it to a master"""
         last = 0  # never have we signed in
         auth_wait = minion.opts["acceptance_wait_time"]
         failed = False
@@ -1154,7 +1185,8 @@ class MinionManager(MinionBase):
                 if minion.opts.get("scheduler_before_connect", False):
                     minion.setup_scheduler(before_connect=True)
                 if minion.opts.get("master_type", "str") != "disable":
-                    yield minion.connect_master(failed=failed)
+                    await minion.connect_master(failed=failed)
+                log.error("RUN MINION TUNE IN")
                 minion.tune_in(start=False)
                 self.minions.append(minion)
                 break
@@ -1164,11 +1196,12 @@ class MinionManager(MinionBase):
                     "Error while bringing up minion for multi-master. Is "
                     "master at %s responding?",
                     minion.opts["master"],
+                    exc_info=True,
                 )
                 last = time.time()
                 if auth_wait < self.max_auth_wait:
                     auth_wait += self.auth_wait
-                yield tornado.gen.sleep(auth_wait)  # TODO: log?
+                await tornado.gen.sleep(auth_wait)  # TODO: log?
             except SaltMasterUnresolvableError:
                 err = (
                     "Master address: '{}' could not be resolved. Invalid or"
@@ -3269,16 +3302,16 @@ class Minion(MinionBase):
             self.pub_channel.on_recv(None)
             log.error("create pub_channel.close task %r", self)
             self.pub_channel.close()
-            #self.io_loop.asyncio_loop.run_until_complete(self.pub_channel.close())
-            #if hasattr(self.pub_channel, "close"):
+            # self.io_loop.asyncio_loop.run_until_complete(self.pub_channel.close())
+            # if hasattr(self.pub_channel, "close"):
             #    asyncio.create_task(
             #        self.pub_channel.close()
             #    )
             #    #self.pub_channel.close()
-            #del self.pub_channel
+            # del self.pub_channel
         if hasattr(self, "event"):
             log.error("HAS EVENT")
-        #if hasattr(self, "periodic_callbacks"):
+        # if hasattr(self, "periodic_callbacks"):
         #    for cb in self.periodic_callbacks.values():
         #        cb.stop()
         log.error("%r destroy method finished", self)
