+++ salt/utils/asynchronous.py
@@ -59,7 +59,9 @@ class SyncWrapper:
         loop_kwarg=None,
     ):
         self.asyncio_loop = asyncio.new_event_loop()
-        self.io_loop = tornado.ioloop.IOLoop(asyncio_loop=self.asyncio_loop)
+        self.io_loop = tornado.ioloop.IOLoop(
+            asyncio_loop=self.asyncio_loop, make_current=False
+        )
         if args is None:
             args = []
         if kwargs is None:
@@ -108,12 +110,16 @@ class SyncWrapper:
                 log.error("No async method %s on object %r", method, self.obj)
             except Exception:  # pylint: disable=broad-except
                 log.exception("Exception encountered while running stop method")
+        # thread = threading.Thread(target=self._run_loop_final, args=(self.asyncio_loop,))
+        # thread.start()
+        # thread.join()
         io_loop = self.io_loop
         io_loop.stop()
         try:
             io_loop.close(all_fds=True)
         except KeyError:
             pass
+        self.asyncio_loop.close()
 
     def __getattr__(self, key):
         if key in self._async_methods:
@@ -137,6 +143,18 @@ class SyncWrapper:
 
         return wrap
 
+    def _run_loop_final(self, asyncio_loop):
+        asyncio.set_event_loop(asyncio_loop)
+        io_loop = tornado.ioloop.IOLoop.current()
+        try:
+
+            async def noop():
+                await asyncio.sleep(0.2)
+
+            result = io_loop.run_sync(lambda: noop())
+        except Exception:  # pylint: disable=broad-except
+            log.error("Error on last loop run")
+
     def _target(self, key, args, kwargs, results, asyncio_loop):
         asyncio.set_event_loop(asyncio_loop)
         io_loop = tornado.ioloop.IOLoop.current()
@@ -149,7 +167,16 @@ class SyncWrapper:
             results.append(sys.exc_info())
 
     def __enter__(self):
+        if hasattr(self.obj, "__aenter__"):
+            ret = self._wrap("__aenter__")()
+            if ret == self.obj:
+                return self
+            else:
+                return ret
         return self
 
     def __exit__(self, exc_type, exc_val, tb):
-        self.close()
+        if hasattr(self.obj, "__aexit__"):
+            return self._wrap("__aexit__")
+        else:
+            self.close()
