+++ tests/support/pytest/transport.py
@@ -73,13 +73,21 @@ class Collector(salt.utils.process.SignalHandlingProcess):
             "rotate_master_key": self._rotate_secrets,
         }
 
+    def _teardown_listener(self):
+        if self.transport == "zeromq":
+            self.sock.close()
+            self.ctx.term()
+        else:
+            self.sock.close()
+
     def _setup_listener(self):
         if self.transport == "zeromq":
-            ctx = zmq.Context()
-            self.sock = ctx.socket(zmq.SUB)
+            self.ctx = zmq.Context()
+            self.sock = self.ctx.socket(zmq.SUB)
             self.sock.setsockopt(zmq.LINGER, -1)
             self.sock.setsockopt(zmq.SUBSCRIBE, b"")
             pub_uri = "tcp://{}:{}".format(self.interface, self.port)
+            log.error("Collector listen %s", pub_uri)
             self.sock.connect(pub_uri)
         else:
             end = time.time() + 120
@@ -97,22 +105,28 @@ class Collector(salt.utils.process.SignalHandlingProcess):
 
     @tornado.gen.coroutine
     def _recv(self):
+        # log.error("RECV %s", self.transport)
         if self.transport == "zeromq":
             # test_zeromq_filtering requires catching the
             # SaltDeserializationError in order to pass.
             try:
                 payload = self.sock.recv(zmq.NOBLOCK)
+                # log.error("ZMQ Payload is %r", payload)
                 serial_payload = salt.payload.loads(payload)
                 raise tornado.gen.Return(serial_payload)
             except (zmq.ZMQError, salt.exceptions.SaltDeserializationError):
                 raise RecvError("ZMQ Error")
         else:
             for msg in self.unpacker:
-                raise tornado.gen.Return(msg["body"])
+                # log.error("TCP Payload is %r", msg)
+                serial_payload = salt.payload.loads(msg["body"])
+                # raise tornado.gen.Return(msg["body"])
+                raise tornado.gen.Return(serial_payload)
             byts = yield self.sock.read_bytes(8096, partial=True)
             self.unpacker.feed(byts)
             for msg in self.unpacker:
-                raise tornado.gen.Return(msg["body"])
+                serial_payload = salt.payload.loads(msg["body"])
+                raise tornado.gen.Return(serial_payload)
             raise RecvError("TCP Error")
 
     @tornado.gen.coroutine
@@ -128,40 +142,44 @@ class Collector(salt.utils.process.SignalHandlingProcess):
         self.start = last_msg
         serial = salt.payload.Serial(self.minion_config)
         crypticle = salt.crypt.Crypticle(self.minion_config, self.aes_key)
-        while True:
-            curr_time = time.time()
-            if time.time() > self.hard_timeout:
-                log.error("Hard timeout reaced in test collector!")
-                break
-            if curr_time - last_msg >= self.timeout:
-                log.error("Receive timeout reaced in test collector!")
-                break
-            try:
-                payload = yield self._recv()
-            except RecvError:
-                time.sleep(0.01)
-            else:
+        try:
+            while True:
+                curr_time = time.time()
+                if time.time() > self.hard_timeout:
+                    log.error("Hard timeout reaced in test collector!")
+                    break
+                if curr_time - last_msg >= self.timeout:
+                    log.error("Receive timeout reaced in test collector!")
+                    break
                 try:
-                    payload = crypticle.loads(payload["load"])
-                    if not payload:
-                        continue
-                    if "start" in payload:
-                        log.info("Collector started")
-                        self.running.set()
-                        continue
-                    if "stop" in payload:
-                        log.info("Collector stopped")
-                        break
-                    last_msg = time.time()
-                    self.results.append(payload["jid"])
-                except salt.exceptions.SaltDeserializationError:
-                    log.error("Deserializer Error")
-                    if not self.zmq_filtering:
-                        log.exception("Failed to deserialize...")
-                        break
-        self.end = time.time()
-        print(f"Total time {self.end - self.start}")
-        loop.stop()
+                    payload = yield self._recv()
+                except RecvError:
+                    time.sleep(0.03)
+                else:
+                    try:
+                        log.trace("Colleted payload %r", payload)
+                        payload = crypticle.loads(payload["load"])
+                        if not payload:
+                            continue
+                        if "start" in payload:
+                            log.info("Collector started")
+                            self.running.set()
+                            continue
+                        if "stop" in payload:
+                            log.info("Collector stopped")
+                            break
+                        last_msg = time.time()
+                        self.results.append(payload["jid"])
+                    except salt.exceptions.SaltDeserializationError:
+                        log.error("Deserializer Error")
+                        if not self.zmq_filtering:
+                            log.exception("Failed to deserialize...")
+                            break
+            self.end = time.time()
+            print(f"Total time {self.end - self.start}")
+        finally:
+            self._teardown_listener()
+            loop.stop()
 
     def run(self):
         """
@@ -173,11 +191,17 @@ class Collector(salt.utils.process.SignalHandlingProcess):
         loop.start()
 
     def __enter__(self):
+        import sys
+
+        print("COL ENTER")
+        sys.stdout.flush()
         self.manager.__enter__()
         self.start()
         # Wait until we can start receiving events
         self.started.wait()
         self.started.clear()
+        print("COL ENTER - Done")
+        sys.stdout.flush()
         return self
 
     def __exit__(self, *args):
@@ -197,6 +221,7 @@ class Collector(salt.utils.process.SignalHandlingProcess):
 class PubServerChannelProcess(salt.utils.process.SignalHandlingProcess):
     def __init__(self, master_config, minion_config, **collector_kwargs):
         super().__init__()
+        self.name = "PubServerChannelProcess"
         self._closing = False
         self.master_config = master_config
         self.minion_config = minion_config
@@ -226,17 +251,24 @@ class PubServerChannelProcess(salt.utils.process.SignalHandlingProcess):
             self.master_config["interface"],
             self.master_config["publish_port"],
             self.aes_key,
-            **self.collector_kwargs
+            **self.collector_kwargs,
         )
 
     def run(self):
+        import queue
+
+        ioloop = tornado.ioloop.IOLoop()
         try:
             while True:
-                payload = self.queue.get()
+                try:
+                    payload = self.queue.get(False)
+                except queue.Empty:
+                    time.sleep(0.03)
+                    continue
                 if payload is None:
                     log.debug("We received the stop sentinel")
                     break
-                self.pub_server_channel.publish(payload)
+                ioloop.run_sync(lambda: self.pub_server_channel.publish(payload))
         except KeyboardInterrupt:
             pass
         finally:
@@ -253,8 +285,8 @@ class PubServerChannelProcess(salt.utils.process.SignalHandlingProcess):
         if self.process_manager is None:
             return
         self.process_manager.terminate()
-        if hasattr(self.pub_server_channel, "pub_close"):
-            self.pub_server_channel.pub_close()
+        if hasattr(self.pub_server_channel, "close"):
+            self.pub_server_channel.close()
         # Really terminate any process still left behind
         for pid in self.process_manager._process_map:
             terminate_process(pid=pid, kill_children=True, slow_stop=False)
@@ -264,9 +296,12 @@ class PubServerChannelProcess(salt.utils.process.SignalHandlingProcess):
         self.queue.put(payload)
 
     def __enter__(self):
+        log.error("Proc start")
         self.start()
+        log.error("Col enter")
         self.collector.__enter__()
         attempts = 300
+        log.error("Wait collector")
         while attempts > 0:
             self.publish({"tgt_type": "glob", "tgt": "*", "jid": -1, "start": True})
             if self.collector.running.wait(1) is True:
@@ -274,6 +309,7 @@ class PubServerChannelProcess(salt.utils.process.SignalHandlingProcess):
             attempts -= 1
         else:
             pytest.fail("Failed to confirm the collector has started")
+        log.error("Collector started")
         return self
 
     def __exit__(self, *args):
@@ -290,4 +326,4 @@ class PubServerChannelProcess(salt.utils.process.SignalHandlingProcess):
         self.stopped.wait(10)
         self.close()
         self.terminate()
-        log.info("The PubServerChannelProcess has terminated")
+        log.error("The PubServerChannelProcess has terminated")
